move switch to skeletal animation component
