+++ cocos/core/3d/framework/skinning-model-component.ts
@@ -28,13 +28,11 @@
  */
 
 import { AnimationClip } from '../../animation/animation-clip';
-import { Material } from '../../assets/material';
 import { Skeleton } from '../../assets/skeleton';
 import { ccclass, executeInEditMode, executionOrder, menu, property } from '../../data/class-decorator';
-import { FlexibleSkinningModel } from '../../renderer/models/flexible-skinning-model';
+import { BakedSkinningModel } from '../../renderer/models/baked-skinning-model';
 import { SkinningModel } from '../../renderer/models/skinning-model';
 import { Node } from '../../scene-graph/node';
-import { builtinResMgr } from '../builtin';
 import { ModelComponent } from './model-component';
 
 /**
@@ -91,66 +89,32 @@ export class SkinningModelComponent extends ModelComponent {
     }
 
     get model () {
-        return this._model as SkinningModel | FlexibleSkinningModel;
+        return this._model as SkinningModel | BakedSkinningModel | null;
     }
 
     constructor () {
         super();
-        this._modelType = SkinningModel;
+        this._modelType = BakedSkinningModel;
     }
 
     public uploadAnimation (clip: AnimationClip | null) {
         this._clip = clip;
-        if (this._model instanceof SkinningModel) {
-            this._model.uploadAnimation(clip);
+        if (this.model && this.model.uploadAnimation) {
+            this.model.uploadAnimation(clip);
         }
     }
 
-    public get realTimePoseCalculation () {
-        const mat = this._materialInstances[0] || this._materials[0];
-        const pass = mat && mat.passes[0];
-        const res = pass && pass.defines.REALTIME_POSE_CALCULATION || false;
-        if (CC_EDITOR) {
-            for (let i = 0; i < this._materials.length; i++) {
-                const m = this._materialInstances[i] || this._materials[i];
-                if (!m) { continue; }
-                for (let j = 0; j < m.passes.length; j++) {
-                    if (res !== (m.passes[j].defines.REALTIME_POSE_CALCULATION || false)) {
-                        console.warn(`${this.node.name}: REALTIME_POSE_CALCULATION should be consistently declared among submodels and passes`);
-                        m.passes[j].defines.REALTIME_POSE_CALCULATION = res; m.passes[j].tryCompile();
-                    }
-                }
-            }
-        }
-        return res as boolean;
-    }
-
-    protected _updateModels () {
-        let modelType = this._modelType;
-        modelType = this.realTimePoseCalculation ? FlexibleSkinningModel : SkinningModel;
+    public updateAnimatingMode (baked = true) {
+        const modelType = baked ? BakedSkinningModel : SkinningModel;
         if (this._model && modelType !== this._modelType) {
             cc.director.root.destroyModel(this._model);
             this._model = null;
             this._models.length = 0;
-        }
-        this._modelType = modelType;
-        super._updateModels();
-    }
-
-    protected _onMaterialModified (idx: number, material: Material | null) {
-        let modelType = this._modelType;
-        modelType = this.realTimePoseCalculation ? FlexibleSkinningModel : SkinningModel;
-        if (modelType !== this._modelType) {
             this._modelType = modelType;
-            if (this._model) {
-                cc.director.root.destroyModel(this._model);
-                this._model = null;
-                this._models.length = 0;
-                super._updateModels();
+            this._updateModels();
+            if (this.enabledInHierarchy) {
                 this._attachToScene();
             }
-        } else {
-            super._onMaterialModified(idx, material);
         }
     }
 
@@ -159,17 +123,10 @@ export class SkinningModelComponent extends ModelComponent {
         super._updateModelParams();
     }
 
-    protected _getBuiltinMaterial () {
-        // classic ugly pink indicating missing material
-        return builtinResMgr.get<Material>(`missing-${this._modelType === SkinningModel ? '' : 'flexible-'}skinning-material`);
-    }
-
     private _update () {
-        if (this._model) {
-            (this._model as SkinningModel | FlexibleSkinningModel).bindSkeleton(this._skeleton, this._skinningRoot, this._mesh);
-            if (this._model instanceof SkinningModel) {
-                this._model.uploadAnimation(this._clip);
-            }
+        if (this.model) {
+            this.model.bindSkeleton(this._skeleton, this._skinningRoot, this._mesh);
+            if (this.model.uploadAnimation) { this.model.uploadAnimation(this._clip); }
         }
     }
 }
